scripts/social_share/requirements.txt
anthropic>=0.39.0
openai>=1.50.0
tweepy>=4.14.0
atproto>=0.0.55
pyyaml>=6.0
tomli>=2.0.0;python_version<"3.11"
